Remove redundant tests on AO/AOCO tables.
We have tons of tests on creating AO and AOCO tables in the regular
regression suite, with and without checksum. And a lot of tests on inserts
to AO/AOCO tests. There are also tests on having indexes on tables,
including AO/AOCO tests. I don't see anything interesting in these
particular combinations.