Disable ORCA also for the internal query to get index size.
For the same reasons we disabled ORCA in all the other ANALYZE queries.
Missed this one.

I belive we don't want to use ORCA for any of the ANALYZE's internal
queries, current or future ones, so move the disabling of ORCA one level
up, into a wrapper around analyze_rel().