Use the new optimizer_trace_fallback GUC to detect ORCA fallbacks.
qp_olap_group2 did an elaborate dance with optimizer_log=on,
client_min_messages=on, and gpdiff rules to detect whether the queries
fell back to the traditional planner. Replace all that with the new simple
optimizer_trace_fallback GUC.

Also enable optimizer_trace_fallback in the 'gp_optimizer' test. Since this
test is all about testing ORCA, seems appropriate to memorize which queries
currently fall back and which do not, so that we detect regressions where
we start to fall back on queries that ORCA used to be able to plan. There
was one existing test that explicitly set client_min_messages=on, like the
tests in qp_olap_group2, to detect fallback. I kept the those extra logging
GUCs for that one case, so that we have some coverage for that too, although
I'm not sure how worthwhile it is anymore.

In the passing, in the one remaining test in gp_optimizer that sets
client_min_messages='log', don't assume that log_statement is set to 'all'.

Setting optimizer_trace_fallback=on for 'gp_optimizer' caught the issue
I fixed in previous commit, that one of the ANALYZE queries still used ORCA.