Use available Makefile variables for src/bin
Since we have a boolean yes/no switch for coverage enabled builds
in Makefile.global from autoconf, use that rather than inspecting
the CFLAGS for ease of reading the code. Also use the top_builddir
variable in setting the libdir, while unlikely to move in the
hierarchy it's the right thing to do.