Multi-dimensional arrays can now be used as arguments to a PL/python … (#1292)
* Multi-dimensional arrays can now be used as arguments to a PL/python function
(used to throw an error), and they can be returned as nested Python lists.

This makes a backwards-incompatible change to the handling of composite
types in arrays. Previously, you could return an array of composite types
as "[[col1, col2], [col1, col2]]", but now that is interpreted as a two-
dimensional array. Composite types in arrays must now be returned as
Python tuples, not lists, to resolve the ambiguity. I.e. "[(col1, col2),
(col1, col2)]".

To avoid breaking backwards-compatibility, when not necessary, () is still
accepted for arrays at the top-level, but it is always treated as a
single-dimensional array. Likewise, [] is still accepted for composite types,
when they are not in an array. Update the documentation to recommend using []
for arrays, and () for composite types, with a mention that those other things
are also accepted in some contexts.

The upstream patch is 94aceed31773

* Give a hint, when [] is incorrectly used for a composite type in array.

    That used to be accepted, so let's try to give a hint to users on why
    their PL/python functions no longer work.

   see upstream 510e1b8ecf

* add tests for additional conversions.
move them into plpython_types
move the create function closer to the actual select
which is more like upstream