Refactor Expr to DXL translator for list partition selector predicates
Previously when the final plan is translated into DXL, Orca uses range
predicates to represent the values of the list partition (range start and end
value equal a single list value), which is error-prone and redundant for the
query executor.

In this patch, we use the following way to translate the predicates of the list
partition selector in DXL (range based partition remain unchanged):
1. PK op Scalar -> Scalar reverse(op) Any(PartListValues)
For partition selector predicate `pk1 < 5`, will be translated to `ArrayComp(5 >
Any(PartListValues))`, which means as long as any value of list partition values
is less than 5, the partition will be selected.
2. PK is (not) NULL -> PartListNullTest
3. Propagation Expression ->
   Const1 = Any(PartListValues) or Const2 = Any(PartListValues) ...

[#140699737]
Closes #149