Serialize minidumps in a streaming fashion.
Instead of having a fixed-size buffer to serialize minidumps to, refactor
the serialization functions to write to a stream. That simplifies the
serialization functions, as they no longer need to reserve space in the
buffer ahead of time (i.e. the UlpRequiredSpace() functions are gone),
reduces memory usage when dumping small queries, and makes it possible to
minidump large queries without running out of memory.

This removes the arbitrary 16 MB limit on minidump size. I'm not sure it's
a good idea to create multi-gigabyte minidumps in practice, but that's a
case of "if it hurts, don't do it". At least it's now possible, if you need
to.