Remove unused BugBuster leftovers
With the last remaining testsuites moved over to ICW, there is no
longer anything left running in BugBuster. Remove the remaining
files and BugBuster makefile integration in one big swing of the
git rm axe. The only thing left in use was a data file which was
referenced from ICW, move this to regress/data instead.