Use appendStringInfoString() where possible
appendStringInfo() is a variadic function treating the passed
string as a format specifier. This is wasteful processing when
just adding a constant string which can be done faster with a
call to appendStringInfoString() where no format processing is
performed.

This leaves lots of appendStringInfo() calls in the processing
but they are from upstream and will be addressed when we merge
with future versions of postgres. The calls in this patch are
the GPDB specific ones.