Update docs and README with new locations
Set the correct Wiki URL to use for the project, we have moved over
to using the Github wiki on the main repo and not the wiki on the
website repo.

Remove references to Pivotal Greenplum documentation in the README
and stick to greenplum.orc/docs/ there as those are the applicable
docs.  While there update the README.md to actually match reality,
the docs are now in the main repo.