Expression dedup complexity from Quad to Linear
This commit introduces a performance improvement and stability for
deduplication process of passed expression list.

Earlier the running time complexity of this operation was O(n^2). This
commit reduces the running time to O(n) using a HashMap.

In addition, previously the order of expression is not preserved, as we
used to add the last entry for a expr in the output list, however now
we insert the first entry which keeps the order as is.

E.g. Let's assume everything in the below list is an expression:

A = `a,a,b,b,c,c`
If you run `dedup(A)` It will produce B=`a,b,c`.

Then if you add another `a` to the list B, it will B'=`a,b,c,a`
and when we run the `dedup(B')` it will generate `b,c,a` in the old
code.

With this change, it will always generate `a,b,c`.

Signed-off-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>