Fixed Test Cases
With improvement in the expression de duplication algorithm, we now
also retain the order of expression due to which the test cases required
the order expression to be fixed.

We see a lot of plan diffs and plan size changes and the reason for that
is the we always preserve the order of deduped list.

Since we output different deduped list the plan changes are expected
specifically the way that we hash join and hash distribute.

Signed-off-by: Bhunvesh Chaudhary <bchaudhary@pivotal.io>