Finish up the removal of "relation of an index" [#143349743]
While working on [#120303669] (greenplum-db/gporca#160) we found that
the abstraction in ORCA contains "the relation of an index". We don't
use this information anywhere, instead, we get indices from "table
descriptors", which contains "indices of a relation", not the other way
around. And retaining this dead abstraction actually introduces undue
complexity for (greenplum-db/gporca#160) and (greenplum-db/gpdb#2163) .
We are removing all mentions of "relation of an index".

This commit removes `RelationMdid` from the dxl:Index nodes from all our
DXL files.