Revert(except test) "[#129402309] Fix double free when using stream bitmaps with hash bitmaps."
This reverts commit b5d7b30fd36787a0cb3a7d28a2dadc9e259deda2.

`HashStreamOpaque` will never own the `HashBitmap` and hence there is no
need of variable `free` in `HashStreamOpaque` to denote if it owns the
`HashBitmap` or not. Hence we don't need `tbm_create_stream_node_ref`
API at all. Given that `HashStreamOpaque` is not owning the `HashBitMap`
it should never try to access the `HashBitMap` during the free of
`HashStreamOpaque`. There is no guarantee that `HashBitmap` is valid at
that time. `ExecEndBitmapIndexScan` might have freed the `HashBitmap`
that is referenced in `HashStreamOpaque`.