Fix collocation checks in windows query planning.
In assure_collocation_and_order, we check if the lower plan is already
distributed on the partitioning key, and insert a Motion node if not.
That check was broken. When a query with multiple windows is planned, we
create a chain of Group+Sort stages, to compute each aggregate. To do that,
we create a synthetic subquery for each stage, and each subquery will have
its own minimal range table and target list. The sort and distribution
path keys should use expressions that refer to the original plan and
PlannerInfo that we started with, but in assure_collocation_and_order(), we
were incorrectly building path keys based on the target lists in the
intermediate stages. Those are not comparable with path keys referring to
the original target list, and as a result.

Fixes github issue #2236.