Fix bfv_partition to avoid explain_format by atmsort [#143465051]
bfv_partition was migrated from TINC, which didn't seem to filter output
through atmsort.pl . This presents us a surprise: a lot of our queries
will trigger an "EXPLAIN re-format" in atmsort where we did NOT intend
it to happen: we already run PL/Python over the EXPLAIN output to count
operators. This surprising behavior also crept up on us when we had a
test failure, and we saw out-of-order output lines in the diff.

This commit fixes that.