Fix issues reported by Coverity
## CID 130007 Resource leak (file descriptor) in XlogFillZero()

Coverity thinks that a file descriptor with value 0 may be leaked (no
matching close() for an open()).  But open() will never return a file
descriptor 0 in this case.  Change the code still to make Coverity happy.

## CID 129351 Dead code in XLogAddRecordsToChangeTracking()

This is a legitimate case.  The `while (1)` loop already scans xlog
until the end, so that (record == NULL) is true.  There is no point in
trying to traverse till the end of xlog yet another time aftre
breaking out of the `while (1)`.  So replace the code and add assert
instead.