Ensure reader gang for subtransaction test is on segments [#138725549]
The sub-transaction tests are injecting faults at a point that only
happens in a reader gang on segments. The original query, when planned
by ORCA, would lead to a plan where the reader slice executes on master.
This commit changes the test (while mostly staying true to its spirits):
instead of using `INSERT INTO ... VALUES ...` we put the value in a
temporary table, and do an `INSERT INTO ... SELECT ... FROM`. The new
test should now pass under both ORCA and planner.