Fix memory leak when using plan cache
When a cached plan is used, this plan resides in a memory context called
CachedPlan. GPDB would make a copy of the plan before dispatching, to evaluate
the STABLE/IMMUTABLE functions, and this copy is allocated in the same memory
context with the cached plan, hence in CachedPlan context. This memory context
would not be freed across queries, so OOM would be hit if plan cache is heavily
used.

Fix is to copy the cached plan into portal's memory context, so the copy
mentioned above would be in portal's memory context as well, and would be freed
when the portal is dropped.