Retire SAN failover strategy
The SAN FTS failover strategy is deprecated and no longer maintained,
as it's no longer needed. Retire the dead code to clean up. Since we
currently are in catalog freeze, the gp_san_configuration catalog is
left in place but will be removed once we open up for the next major
version cycle. Since the gp_san_config header file contains catalogs
for general FTS fault strategies, rename the file for clarity.