Move gp_workfile_mgr to internal_tools contrib
The gp_workfile_mgr functions are installed via gp_toolkit, so
being in src/bin is the wrong place for such code. Move to the
gp_internal_tools contrib, which also is the wrong place, but
at least has precedence for containing gp_toolkit functions.

These should at some point be made into backend functions and
moved from contrib but not at this point in cycle.