Coverity fixes for xlog_mm.c
CID 129288, 129277:
Buffer not null terminated (BUFFER_SIZE_WARNING)
buffer_size_warning: Calling strncpy with a maximum size argument of 1024 bytes
on destination array m->path2 of size 1024 bytes might leave the destination
string unterminated.
This issue was detected in the `xlog_mm_redo()` function. Fixed by using
`strlcpy()` to ensure that the buffer is null terminated.

CID 129856: Argument cannot be negative (NEGATIVE_RETURNS)
negative_returns: fd is passed to a parameter that cannot be negative.
This issue was also found in the `xlog_mm_redo()` function.
Fixed the issue by making sure that `close()` is not called on negative fd.