Properly get smallest eof column during ADD COLUMN on AOCO table
The column_to_scan function used to get the smallest eof column was
operating incorrectly. It would loop the segfiles until it found a
segfile with nonzero tuplecount, grab the first column that has a
nonzero eof, and return that value. The issue here is that it did not
iterate through all the columns to find the smallest nonzero eof. With
this commit, we now loop through all the columns of the segfile to
decide on a column to return. The unit tests for this function were
also incorrect so we fix that in this commit too.