Use a different variable for gpaddon git access key
This change was made in order to facilitate the new git branching rules.
In order to use different credentials to pull in from our fork, we need
to be able to more finely control which git repo we are pulling from
when we clone the pipeline to smoke test a change.

Signed-off-by: Divya Bhargov <dbhargov@pivotal.io>