Add binary compatibility testing.
We need some tests to see if binaries can be swapped around. This
commit provides a simple framework and basic tests to check if
binaries can be upgraded/downgraded by just swapping the Greenplum
binary install paths. Running the tests requires the user to have
another Greenplum compiled which will be provided as an argument to
the test script. It also requires the regression database to be built
from running tests in src/test/regress.