Sync JSON functions with PostgreSQL 9.3.17.
We had cherry-picked some of the PostgreSQL 9.3 functions, but not all the
subsequent fixes to them. I believe we need at least upstream commit
66008564f8 to fix github issue #2430, and surely the other bug fixes are
needed too.

Rather than cherry-pick individual commits, this replaces the JSON related
source files with the upstream files in toto. This will make diffing and
maintenance of these backported JSON function easier in the future.