Add binary compatibility tests to Concourse pipeline.
This is needed to test that Greenplum HEAD is binary compatible with a
previous version of Greenplum. This is controlled by providing a
branch/tag to the gpdb_src_binary_swap git resource. We run the tests
in the same job where we run ICW to reuse the regression database
generated. To start this off, we'll be testing that HEAD can binary
swap over 5.0.0-alpha.5.