Use unittest module instead of unittest2 module in python unittests.
Since python 2.7 has the unittest module included, we do not need to use
unittest2 anymore. This removes this dependency requirement. We
previously used the unittest2 module as it included features that were
not present in python 2.6. There are still unittest2 dependencies in the
TINC tests, but that may be a more intensive effort.