Fix support for dumping ProcLangs with anonymous block support
The support for anonymous blocks which was backported in commit
cacb2839ac869b3f0ab9e9fc4b152f7d83e71 missed updatint the query
in getProcLangs() to correctly extract the laninline function.
Fix by updating the query to what it looks like in upstream 9.0
where anonymous blocks first appeared.

Added a small comment to explain why for when this conflicts in
a merge. There shouldn't be a conflict until we merge 9.0 since
getProcLangs() was quite stable during 8.4, but one never know.