Change name of class
This test class was named the same, "GpTestCase", as the name of the
superclass we use for all tests. Since the packaging was different, this
was not a python conflict, but it was odd and misleading. This class can
be named anything, so it seems easy to change it.

Also modernized the superclass, optimized imports, and added a main()

Signed-off-by: Larry Hamel <lhamel@pivotal.io>