Extend debugging output in missing Oid assignment
When a pre-assigned Oid can't be found, it's rather helpful to see
the full searchkey that was used rather than just the objname since
some keys lack objname (attrdef's for example). Having hacked this
in multiple times I figured we might as well extend the elog() with
the relevant information.