Remote sub_transaction_limit_removal TINC tests.
Now that subtransaction information is no longer maintained in shared
snapshot between readers and a writer, the tests are obsolete.  Faults
used by the tests are removed as well.

This commit contains a gpdb_master pipeline change - to remove
"subtransaction" job that used to run the tests being removed by this
commit.

Signed-off-by: Asim R P <apraveen@pivotal.io>