Make Concourse pipeline forkable after binary swap job added
The compilation job for the binary swap test portion of the pipeline
used the same remote as our normal compilation job. If someone forked
the pipeline, they may encounter issues if that tag or branch is
missing or not updated in their forked remote of Greenplum. Adding an
additional separate remote variable fixes this.