Fix deadlock on master with ENTRY_DB_SINGLETON
Original deadlock is caused by a reader waiting on a lock which is already held
by the writer of the same MPP session, and another session is waiting for a
conflicting mode on the same lock.

The fix is to avoid checking waitMask conflicts for reader (i.e. MyProc is
different from lockHolderProcPtr).

Detailed discussion of the deadlock issue is at:
https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/OS1-ODIK0P4/ZIzayBbMBwAJ

Two isloation2 tests are added.  One to validate the deadlock does not occur
and another to ensure that granting locks to readers does not starve existing
waiters.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>