Fix the walrep test to validate generation of AO XLOG.
The generate_ao_xlog/generate_aoco_xlog test would fail if there was a page
header or a continuation record in the XLOG sent by the sender since the
validation did not take them into consideration. This patch fixes the
validation of the received XLOG by detecting the page headers and continuation
records and making sure that it skips to the correct record boundary.

Signed-off-by: Abhijit Subramanya <asubramanya@pivotal.io>