Add PID suffix to the output filenames of concurrent tests.
The storage/access_methods/appendonly_eof/test_ao_eof.py test achieves its
concurrency by inheriting from the SQLConcurrencyTestCase class defined in
/tincrepo/mpp/models/sql_concurrency_tc.py.  The run_test() method of this class
is called concurrently from multiple processes. To have a unique filename for
the output of each concurrent invocation, the test uses a timestamp suffix to
the output file. Even though the microsecond part of the timestamp is used, it
is possible that two concurrent runs of the test would get the same microsecond
part for the `datetime.datetime.now()`.  Hence, it is possible that two tests
would have the same timestamp resulting in identical output file names.

The proposed fix is to add the pid also to the timestamp.  That should make it
completely unique.