Update Join Cardinality Estimation for Text/bpchar/varchar/char columns
Histogram intersection depends on the value of the bucket boundaries.
For datatypes like text, varchar, etc. Orca currently uses a hash function
to mark bucket boundaries. This function is slightly useful for equality
with singleton buckets but nothing more. So the previous join stats computation
based on histogram intersection is totally bogus. In this CL, we now modified
it into a NDV (number of distinct values) based estimation.