Fix binary swap test when pg_dump is changed
Currently, our binary swap test will report a diff when pg_dump output
is changed.  This is unintended.  To prevent this, we resource the
latest greenplum_path.sh after starting up Greenplum with the old
greenplum_path.sh.  This allows us to use the latest pg_dumpall
against a cluster running off of the old binaries.  If the tests or
SQL in the binary swap test becomes more complex, this workaround may
need to be replaced.