Delete files on mirror for AO/CO tables.
With --enable-segwalrep, mirror leverages replay of xl_mm_fs_obj records to
delete files. Code was not correctly handling appendonly tables as was calling
`smgrdounlink()`, which is for heap tables or indexes. For AO/CO tables, need to
perform drop of specific single file mentioned in xlog record, which is
performed by `MirroredAppendOnly_Drop()`. Without this code currently files like
<relfilenode>.127, <relfilenode>.129,... etc get left behind on mirror. The
problem was not seen so far as master never stores data for AO/CO tables hence
these files are not created on master. Only when now we start enabling wal
replication for segments this is required.