Add tool to assist in managing WAL replication mirror segments
This gpsegwalrep tool is meant for assisting in segment WAL
replication development which is why it is not placed in
gpMgmt/bin/. The tool is used to initialize, start, stop, and destroy
WAL replication mirror segments. It can also be said that this tool is
a rough example of what a segment walrep tool would look like in
Greenplum 6.x.