Revert "gpperfmon: refactor to make 'skew' calculation more readable"
This code is not correctly recording the data. There seems to be a bug
with how we're modifying the values. Most likely, that we're trying to
change the address instead of the actual value it's pointing to.

We will need to fully run the test this part of the code again and create a new PR.

This reverts commit 411d3c8275e3cae836526a088df26db14690ee8a.