Fix incorrect struct member in backport
The backport of the data checksum catalog changes backported the
relevant GUC from a version which has struct config_bool defined
differently than GPDB. The reason an extra NULL in the config_bool
array initialization wasn't causing a compilation failure is that
there is an extra bool member at the end which is only set during
runtime, reset_val. The extra NULL was "overflowing" into this
member and thus only raised a warning under -Wint-conversion:

    guc.c:1180:15: warning: incompatible pointer to integer
	               conversion initializing 'bool' (aka 'char')
				   with an expression of type 'void *’

Fix by removing the superflous NULL. Since it was setting reset_val
to NULL (and for a GUC which is yet to "do something") there should
be no effects by this.