Relax relation matching logic to handle toast tables
The relation matching logic durin upgrades was very strict, which
caused issues when for example a relation had a toast attribute
which was subsequently dropped. In the new cluster there will be
no toast table for this table. Also don't treat the existence of
new toast tables as a fatal error since newer versions are free
to create toasts where previous versions didn't.

This is a partial backport of upstream commit 73b9952e8278b9c9a6.