Fix gpcheckmirrorseg.pl for MAC
gpcheckmirrorseg.pl assumed GPDB is compiled as 32 bit on macs and hence doesn't
work correctly. Fixing the same as now we compile GPDB in 64 bits on MACs too.

Also, add errlog directory to ignore_list as its not supposed to be same on
primary and mirror. Plus, reflect missing relkinds for visimap, blockdirectory
and sequence tables.