Reformat code.
There was a mixture of spaces and tabs being used for indentation in
aocsam.c, and I finally got fed up with that while doing other changes in
that file. I ran pgindent, and did a bunch of manual fixups of the
formatting. All the changes in this commit are purely cosmetic.

I did the same for appendonlyam.c, although I'm not changing it at the
moment, to keep aocsam.c and appendonlyam.c in sync.